sharp debugging

+++ Projet/Développez_le_back-end d_un_site_de_notation_de_livres/backend/middleware/sharp-conf.js
@@ -1,24 +1,14 @@
-// const express = require("express")
+const sharp = require("sharp")
 const fs = require("fs")
 
-// const multer = require("multer");
-const sharp = require("sharp")
-// const sharp = require('sharp-multer')
-
-// const app = express()
-// const storage = multer.memoryStorage()
-// const upload = multer({ storage })
-
 module.exports = async (req, res, next) => {
-    // if (!req.file) { 
-    //     return next() }
+    if (!req.file) { return next() }
         try {
-            // await sharp(req.file.path)
-            // .webp({ quality: 90 })
-            // .resize({ width: 400, height: 500 })
-            // .toFile(`${ req.file.path.split('.')[0] }transformed.webp`)
-            
-            console.log('Youpi')
+            await sharp(req.file.path)
+                .webp({ quality: 90 })
+                .resize({ width: 400, height: 500 })
+                .toFile(`${ req.file.path.split('.')[0] }transformed.webp`)
+
             fs.unlink(req.file.path, (error) => {
                 req.file.path = `${req.file.path.split('.')[0]}transformed.webp`
                 if (error) { console.log(error, 'Image not deleted') }
